入力された内容に応じて単位変換を行う処理実装

+++ 007_UnitConvertTool/src/main/java/com/example/template/entity/UnitDetailEntity.java
@@ -1,12 +1,14 @@
 package com.example.template.entity;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class UnitDetailEntity {
 
 	private Integer unitId;
-	private Integer unitDetailId;
+	private BigDecimal unitDetailId;
 	private String unitDetailName;
 	private String unitDetailValue;
 }
